Breaking change! Now New() accepts a list of roots.
This allows partial themes or different roots to be used depending on
their position in the list. All roots are searched for the executed
template and the first found is used.
A ne method MustCompile was implemented. It panics instead of returning
error.

Added method Must() - a convenient wrapper for New.

must